Change sub-packages structure to uv extras approach. Update streamlit css loader for new versions.
