+++ {{ cookiecutter.repo_name }}/.pre-commit-config.yaml
@@ -9,6 +9,6 @@ repos:
     # Run the formatter.
     - id: ruff-format
 - repo: https://github.com/kynan/nbstripout
-  rev: 0.5.0
+  rev: 0.6.1
   hooks:
     - id: nbstripout
