+++ {{ cookiecutter.repo_name }}/azureml/azureml_config.yaml
