+++ {{ cookiecutter.repo_name }}/azureml/create_environment.py
