+++ {{ cookiecutter.repo_name }}/azureml/process_raw_dummy.py
