--- {{ cookiecutter.repo_name }}/packages/app_fastapi/pyproject.toml
@@ -1,16 +1,0 @@
-[project]
-name = "app_fastapi"
-version = "0.1.0"
-requires-python = ">={{ cookiecutter.minimum_python_version }}"
-description = "A FastAPI app for {{ cookiecutter.project_name }}"
-authors = [{ name = "{{ cookiecutter.author_name }}" }]
-dependencies = [
-    "fastapi", 
-    "uvicorn", 
-    "pydantic", 
-    "python-dotenv"
-]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
