--- {{ cookiecutter.repo_name }}/packages/app_streamlit/pyproject.toml
@@ -1,16 +1,0 @@
-[project]
-name = "app_streamlit"
-version = "0.1.0"
-requires-python = ">={{ cookiecutter.minimum_python_version }}"
-description = "A Streamlit app for {{ cookiecutter.project_name }}"
-authors = [{ name = "{{ cookiecutter.author_name }}" }]
-dependencies = [
-    "numpy", 
-    "plotly", 
-    "plotly-express", 
-    "streamlit==1.35.0"
-]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
