--- {{ cookiecutter.repo_name }}/packages/azureml/pyproject.toml
@@ -1,14 +1,0 @@
-[project]
-name = "azureml"
-version = "0.1.0"
-requires-python = ">={{ cookiecutter.minimum_python_version }}"
-description = "An azureml package for {{ cookiecutter.project_name }}" 
-authors = [{ name = "{{ cookiecutter.author_name }}" }]
-dependencies = [
-  "azure-ai-ml",
-  "azure-identity",
-]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
